server/requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
transformers
pillow
prometheus-client
python-json-logger
